Some refactor of the Messenger class:
Added a "split" method to the Messenger to split the MPI communicator.  

Added a "_parent" member to the Messenger to keep hold of the potentially split Messenger's parent.  

Removed the "sum" and "max" methods in favor of a "reduce" method that can apply other reduction operators in addition to "sum" and "max".  

Simplified the "print_*" methods to just one method (prinfo) with an optional parameter.  Now, there is only one print method called "prinfo", short for "print information", that takes an optional parameter "all".  If "all" is True, then it prints from all ranks.  Otherwise, if False, it only prints from the master rank.

Also, cleaned up some Messenger serial unit tests to match these changes.