Adding tests to compare output from Serial Messenger and 1-rank Parallel Messenger, to make sure the results are the same.  The point of the Messenger is to hide any need for references to MPI code, making the serial code execute with the same operations as the parallel code.
Includes a fix to the serial partition algorithm, which did not deal with sorting the data by weights since the return list in serial is the same as the list given.  However, to preserve the same order of the return list in serial and parallel, the serial code must sort the input list, too.