Push/pull mechanism is now in place with tests.
I think there is a problem with this interface.  When I initially designed the interface, I assumed that gather/scatter was all that would be needed from the messenger.  With that design, the push/pull methods fit nicely into the same framework, with push/pull corresponding to gather/scatter except "vertical" in the tree of ranks.

With the addition of all of the other collective operations, just as reduce and broadcast, and the point-to-point communication method (sendrecv), this is now confusing.  I think a redesign needs to be done.